+++ afacinemas_scraper/__init__.py
@@ -1,3 +1,3 @@
-from afacinemas_scraper.scraper import Scraper
+from afacinemas_scraper.main import Scraper
 
 __all__ = ["Scraper"]
