+++ tests/test_afacinemas_scraper.py
@@ -1,5 +1,2 @@
-from afacinemas_scraper.__version__ import __version__
-
-
-def test_version():
-    assert __version__ == "0.2.0"
+def test_assert_true():
+    assert True
